+++ src/Unchase.Swashbuckle.AspNetCore.Extensions/Filters/XEnumNamesParameterFilter.cs
@@ -1,18 +1,24 @@
 ﻿using System;
-using System.ComponentModel;
+using System.Collections.Generic;
+using System.IO;
 using System.Linq;
+using System.Xml.XPath;
+using Microsoft.Extensions.Options;
 using Microsoft.OpenApi.Any;
 using Microsoft.OpenApi.Models;
 using Swashbuckle.AspNetCore.SwaggerGen;
 using Unchase.Swashbuckle.AspNetCore.Extensions.Extensions;
+using Unchase.Swashbuckle.AspNetCore.Extensions.Options;
 
 namespace Unchase.Swashbuckle.AspNetCore.Extensions.Filters
 {
-    public class XEnumNamesParameterFilter : IParameterFilter
+    internal class XEnumNamesParameterFilter : IParameterFilter
     {
         #region Fields
 
         private readonly bool _includeXEnumDescriptions;
+        private readonly bool _applyFiler;
+        private readonly HashSet<XPathNavigator> _xmlNavigators = new HashSet<XPathNavigator>();
 
         #endregion
 
@@ -21,18 +27,39 @@ namespace Unchase.Swashbuckle.AspNetCore.Extensions.Filters
         /// <summary>
         /// Constructor.
         /// </summary>
-        /// <param name="includeXEnumDescriptions">If true - add "x-enumDescriptions" extensions from <see cref="DescriptionAttribute"/>.</param>
-        public XEnumNamesParameterFilter(bool includeXEnumDescriptions = false)
+        /// <param name="options"><see cref="FixEnumsOptions"/>.</param>
+        /// <param name="configureOptions">An <see cref="Action{FixEnumsOptions}"/> to configure options for filter.</param>
+        public XEnumNamesParameterFilter(IOptions<FixEnumsOptions> options, Action<FixEnumsOptions> configureOptions = null)
         {
-            _includeXEnumDescriptions = includeXEnumDescriptions;
+            if (options.Value != null)
+            {
+                configureOptions?.Invoke(options.Value);
+                this._includeXEnumDescriptions = options.Value?.IncludeDescriptions ?? false;
+                this._applyFiler = options.Value?.ApplyParameterFilter ?? false;
+                foreach (var filePath in options.Value?.IncludedXmlCommentsPaths)
+                {
+                    if (File.Exists(filePath))
+                    {
+                        this._xmlNavigators.Add(new XPathDocument(filePath).CreateNavigator());
+                    }
+                }
+            }
         }
 
         #endregion
 
         #region Methods
 
+        /// <summary>
+        /// Apply the filter.
+        /// </summary>
+        /// <param name="parameter"><see cref="OpenApiParameter"/>.</param>
+        /// <param name="context"><see cref="ParameterFilterContext"/>.</param>
         public void Apply(OpenApiParameter parameter, ParameterFilterContext context)
         {
+            if (!this._applyFiler)
+                return;
+
             var typeInfo = context.ParameterInfo?.ParameterType ?? context.PropertyInfo.PropertyType;
             var enumsArray = new OpenApiArray();
             var enumsDescriptionsArray = new OpenApiArray();
@@ -47,7 +74,7 @@ namespace Unchase.Swashbuckle.AspNetCore.Extensions.Filters
 
                 if (this._includeXEnumDescriptions)
                 {
-                    enumsDescriptionsArray.AddRange(EnumTypeExtensions.GetEnumValuesDescription(typeInfo));
+                    enumsDescriptionsArray.AddRange(EnumTypeExtensions.GetEnumValuesDescription(typeInfo, this._xmlNavigators));
                     if (!parameter.Extensions.ContainsKey("x-enumDescriptions") && enumsDescriptionsArray.Any())
                     {
                         parameter.Extensions.Add("x-enumDescriptions", enumsDescriptionsArray);
@@ -69,7 +96,7 @@ namespace Unchase.Swashbuckle.AspNetCore.Extensions.Filters
 
                         if (this._includeXEnumDescriptions)
                         {
-                            enumsDescriptionsArray.AddRange(EnumTypeExtensions.GetEnumValuesDescription(genericArgumentType));
+                            enumsDescriptionsArray.AddRange(EnumTypeExtensions.GetEnumValuesDescription(genericArgumentType, this._xmlNavigators));
                             if (!parameter.Extensions.ContainsKey("x-enumDescriptions") && enumsDescriptionsArray.Any())
                             {
                                 parameter.Extensions.Add("x-enumDescriptions", enumsDescriptionsArray);